🧹 Tidy: 六角形の頂点計算ロジックをユーティリティとして抽出・共通化
JarCanvas.tsxとJarShelf.tsxで重複していた六角形の頂点座標計算（hexVertices, miniHexVertices）およびSVG用ポリゴンポイント計算（hexPolygonPoints）を、新設した lib/hexagonUtils.ts に抽出し、各コンポーネントからインポートして使用するように修正しました。
