+++ .jules/tidy.md
@@ -8,3 +8,7 @@
 ## 2026-03-21 - チャートタイトルの共通化とインライン条件分岐の排除
 **学び:** `FeedingChart`、`DiaperChart`、`SleepChart` といった各チャートコンポーネントで、現在選択されているビューに基づくタイトル文字列（「7日間の推移」「生活リズム（過去7日）」「今日 vs 7日平均」）の決定ロジックが、ネストされた三項演算子としてインラインでハードコードされていました。また、Vitestで `useIsMobile` のように `window.matchMedia` を使うコンポーネントをテストする際は、`vitest.config.ts` の `setupFiles` にモックを追加しないとエラーになります。
 **アクション:** ビューごとのタイトルを返すユーティリティ関数 `getChartViewTitle` を `ChartViewToggle.tsx` に切り出し、各チャートからインポートして利用する形に修正しました。また、Vitestのエラーに対処するため `window.matchMedia` のモックをセットアップに追加しました。次回以降も、複数の場所でコピー＆ペーストされている長い条件分岐を見つけた際は、関連するUI要素の近くのユーティリティとして切り出すことを検討します。
+
+## $(date +%Y-%m-%d) - 六角形描画ロジックの共通化
+**学び:** `JarCanvas.tsx`と`JarShelf.tsx`で、六角形の頂点やSVGポリゴンポイントを計算する数式（`hexVertices`, `miniHexVertices`, `hexPolygonPoints`）が個別に実装されており、DRY原則に反していた。
+**アクション:** 独立した数式関数を`lib/hexagonUtils.ts`に抽出し、インポートして再利用する構造に変更。今後のコンポーネントでも同様の計算ロジックを必要とする場合はこのユーティリティを活用する。
